+++ opensearchapi/api_msearch.go
@@ -70,7 +70,7 @@ type MSearchResp struct {
 				Value    int    `json:"value"`
 				Relation string `json:"relation"`
 			} `json:"total"`
-			MaxScore float32     `json:"max_score"`
+			MaxScore *float32    `json:"max_score"`
 			Hits     []SearchHit `json:"hits"`
 		} `json:"hits"`
 		Status       int             `json:"status"`