+++ opensearchapi/api_scroll-get.go
@@ -43,7 +43,7 @@ type ScrollGetResp struct {
 			Value    int    `json:"value"`
 			Relation string `json:"relation"`
 		} `json:"total"`
-		MaxScore float32     `json:"max_score"`
+		MaxScore *float32    `json:"max_score"`
 		Hits     []SearchHit `json:"hits"`
 	} `json:"hits"`
 	ScrollID        *string  `json:"_scroll_id,omitempty"`